Move YouTube Music search to its own class
